% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zplaettes.R
\name{zpalettes}
\alias{zpalettes}
\title{Zealot Palettes}
\usage{
zpalettes()
}
\description{
Quick access to color palettes of professional sports clubs,
University/Colleges, movies, etc.
}
\examples{

# get a list of the available palettes
zpalettes()

}
