% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zplaettes.R
\name{nhl_palettes}
\alias{nhl_palettes}
\title{NHL Palettes}
\usage{
nhl_palettes(name = NULL, active = NULL, hex_only = TRUE)
}
\arguments{
\item{name}{character string to (fuzzy) match to the team names. If
\code{NULL} (default) all teams are returned.}

\item{active}{if \code{TRUE} return only active color palettes, if
\code{FALSE} then return only retired color palettess.  If \code{NULL}
(default) return both active and retired color palettes.}

\item{hex_only}{If \code{TRUE} and the combination of \code{name} and
\code{active} return only one match then return a named character vector for
the color palette hex codes.  Otherwise, return a \code{data.frame}.}
}
\description{
Active and retired color palettes for current and former teams of the
National Hockey League.
}
\examples{

# All NHL palettes:
nhl_palettes()

# Return all active palettes
nhl_palettes(active = TRUE)

# Get the palettes for the Colorado Avalanche
nhl_palettes(name = "Colorado Avalanche")  # Using Full Team Name
nhl_palettes(name = "Avs")                 # Using fuzzy matching

# Get the active palette for the Colorado Avalanche
nhl_palettes(name = "Avs", active = TRUE)
nhl_palettes(name = "Avs", active = TRUE, hex_only = FALSE)

}
